changes in base datasets: now DatasetBase is general one and SegmentationBase and DepthBase inherit it
