+++ depth/depth_utils.py
@@ -7,24 +7,23 @@ from .. import dataset_utils
 
 module_path = os.path.dirname(os.path.realpath(__file__))
 
-class DepthBase(Dataset):
+class DepthBase(dataset_utils.DatasetBase):
 
-    def __init__(self):
-        pass
-
-    def __len__(self):
-        pass
-
-    def __getitem__(self, index):
-        pass
-
-    @abstractmethod
     def load_image(self, filename):
         pass
 
-    @abstractmethod
     def load_depth(self, filename):
         pass
+
+    def generate_load_functions(self):
+        return [self.load_image, self.load_depth]
+
+    # TODO GET RID OF THIS 
+    def base_path(self):
+        return module_path
+
+    def keys(self):
+        return ['image', 'depth']
 
 
 class DepthHDF5(dataset_utils.HDF5Dataset):
